refactor: create week 리팩토링
